+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Do not use useMemo/useCallback as a semantic guarantee
 - A naive useMemoSafe for axios config
 - Fix null types
+- short-hand hooks: useFetch, useAxios
 
 ## [0.7.0] - 2019-01-22
 ### Added